move minmax to bitpacker
move minmax to bitpacker
use minmax in blocked bitpacker
